nlp/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
spacy==3.7.2
langdetect==1.0.9
pydantic==2.5.0
requests==2.31.0
langchain==0.1.20
langchain-community==0.0.38
langchain-text-splitters==0.0.1
langchain-core==0.1.52
langchain-groq==0.1.3
chromadb==0.4.22
sentence-transformers==2.7.0
pypdf==3.17.4
python-docx==1.1.0
python-multipart==0.0.6
numpy==1.24.3
huggingface-hub==0.23.0
groq==0.5.0
python-dotenv==1.0.1